+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+## v1.x.x (TBD)
+* Add support of distributed training to S3IterableDataset
+
+### Breaking changes
+* No breaking changes. 
+
 ## v1.2.7 (October 29, 2024)
 
 ### New features
 * Add support for CRT retries (awslabs/mountpoint-s3#1069).
 * Add support for `CopyObject` API (#242).
+
+### Breaking changes
+* No breaking changes.
 
 ## v1.2.6 (October 9, 2024)